requirements.txt
# Core Backend
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
psycopg[binary]==3.2.3

# Database
alembic==1.14.0
asyncpg==0.30.0  # ← Async PostgreSQL driver
aiosqlite==0.20.0  # ← For SQLite tests

# WebSocket
python-socketio==5.11.4
websockets==13.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Development
ruff==0.7.4
black==24.10.0
isort==5.13.2
